backend/pyproject.toml
[tool.poetry]
name = "nail-art-backend"
version = "0.1.0"
description = "FastAPI backend for nail art visual similarity search"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pillow = "^10.1.0"
numpy = "^1.24.3"
requests = "^2.31.0"
torch = "^2.0.0"
transformers = "^4.35.0"
faiss-cpu = "^1.7.4"
opencv-python = "4.8.1.78"
scipy = "^1.16.1"
pinecone = "^7.3.0"
supabase = "^2.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311" 